# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online hautarzt API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/enquires": {
            "post": {
                "tags": [
                    "Enquires"
                ],
                "summary": "Create a new enquire",
                "description": "Create a new enquire",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "gender",
                                    "date_of_birth",
                                    "phone_number",
                                    "email",
                                    "doctor_id",
                                    "answers"
                                ],
                                "properties": {
                                    "first_name": {
                                        "description": "An enquire customer's first name",
                                        "format": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "description": "An enquire customer's last name",
                                        "format": "string",
                                        "example": "Carter"
                                    },
                                    "email": {
                                        "description": "An enquire customer's e-mail",
                                        "format": "string",
                                        "example": "test@gmail.com"
                                    },
                                    "gender": {
                                        "description": "An enquire customer's gender. Possible values: MALE, FEMALE",
                                        "format": "string",
                                        "example": "MALE"
                                    },
                                    "date_of_birth": {
                                        "description": "An enquire customer's date of birth",
                                        "format": "string",
                                        "example": "1993-02-05"
                                    },
                                    "phone_number": {
                                        "description": "An enquire customer's phone number",
                                        "format": "string",
                                        "example": "+38 (099) 548-54-85"
                                    },
                                    "doctor_id": {
                                        "description": "An enquire's doctor ID",
                                        "format": "string",
                                        "example": "5"
                                    },
                                    "answers": {
                                        "description": "An enquire's answers",
                                        "format": "array"
                                    },
                                    "country": {
                                        "title": "A location's country",
                                        "description": "A location's country",
                                        "format": "string",
                                        "example": "USA"
                                    },
                                    "city": {
                                        "title": "A location's city",
                                        "description": "A location's city",
                                        "format": "string",
                                        "example": "New York"
                                    },
                                    "state": {
                                        "title": "A location's state",
                                        "description": "A location's state",
                                        "format": "string",
                                        "example": "New York"
                                    },
                                    "postal_code": {
                                        "title": "A location's postal code",
                                        "description": "A location's postal code",
                                        "format": "string",
                                        "example": "12345"
                                    },
                                    "address": {
                                        "title": "A location's address",
                                        "description": "A location's address",
                                        "format": "string",
                                        "example": "address"
                                    },
                                    "latitude": {
                                        "title": "A location's latitude",
                                        "description": "A location's latitude",
                                        "format": "double",
                                        "example": 5.123,
                                        "nullable": true
                                    },
                                    "longitude": {
                                        "title": "A location's longitude",
                                        "description": "A location's longitude",
                                        "format": "double",
                                        "example": 8.123,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "An enquire has been succesfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EnquireResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "There are some validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "schema": "ValidationError",
                                    "title": "Validation error",
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email field is required."
                                                    }
                                                },
                                                "first_name": {
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The first name field is required."
                                                    }
                                                }
                                            },
                                            "type": "object",
                                            "format": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal technical error was happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "Server Error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/enquires/{id}": {
            "get": {
                "tags": [
                    "Enquires"
                ],
                "summary": "Get an enquires resource by id",
                "description": "Get a enquires resource by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "An enquire's identificator",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An enquire has been succesfully received",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EnquireResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authorization failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Current user has not permissions to do this action",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "No query results for model [App\\Models\\Enquire]."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal technical error was happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "format": "string",
                                            "example": "Something went wrong, please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CarbonResource": {
                "title": "Carbon resource",
                "description": "Resource for datetime representation",
                "properties": {
                    "date": {
                        "format": "string",
                        "example": "2019-12-11 12:25:00.000000"
                    },
                    "timezone_type": {
                        "format": "int64",
                        "example": 3
                    },
                    "timezone": {
                        "format": "string",
                        "example": "Europe/Berlin"
                    }
                },
                "type": "object"
            },
...
            "EnquireResource": {
                "title": "Enquire resource",
                "description": "Resource for an enquire representation",
                "properties": {
                    "id": {
                        "schema": "Enquire",
                        "title": "ID",
                        "description": "An enquire's identificator",
                        "format": "int64",
                        "example": "5"
                    },
                    "first_name": {
                        "schema": "Enquire",
                        "title": "First name",
                        "description": "An enquire customer's first name",
                        "format": "string",
                        "example": "Davide"
                    },
                    "last_name": {
                        "schema": "Enquire",
                        "title": "Last name",
                        "description": "An enquire customers's last name",
                        "format": "string",
                        "example": "Donghi"
                    },
                    "gender": {
                        "schema": "Enquire",
                        "title": "Gender",
                        "description": "An enquire customer's gender",
                        "format": "string",
                        "example": "MALE"
                    },
                    "date_of_birth": {
                        "schema": "Enquire",
                        "title": "Date of birth",
                        "description": "Date of birth representation",
                        "format": "object",
                        "$ref": "#/components/schemas/CarbonResource"
                    },
                    "created_at": {
                        "schema": "Enquire",
                        "title": "Created at",
                        "description": "Created datetime representation",
                        "format": "object",
                        "$ref": "#/components/schemas/CarbonResource"
                    },
                    "updated_at": {
                        "schema": "Enquire",
                        "title": "Updated at",
                        "description": "Last updated datetime representation",
                        "format": "object",
                        "$ref": "#/components/schemas/CarbonResource"
                    },
                    "phone_number": {
                        "schema": "Enquire",
                        "title": "Phone number",
                        "description": "An enquire customer's phone number",
                        "format": "string",
                        "example": "+38 024-548-58-55"
                    },
                    "email": {
                        "schema": "Enquire",
                        "title": "E-mail",
                        "description": "An enquire customer's e-mail",
                        "format": "string",
                        "example": "test@gmail.com"
                    },
                    "conclusion": {
                        "schema": "Enquire",
                        "title": "Conclusion",
                        "description": "A doctors's conclusion",
                        "format": "string",
                        "example": "Everythink will be fine"
                    },
                    "conclusion_created_at": {
                        "schema": "Enquire",
                        "title": "Conclusion created at",
                        "description": "Conclusion created datetime representation",
                        "format": "object",
                        "$ref": "#/components/schemas/CarbonResource"
                    },
                    "status": {
                        "schema": "Enquire",
                        "title": "Status",
                        "description": "An enquire's status",
                        "format": "string",
                        "example": "UNREAD"
                    },
                    "is_paid": {
                        "schema": "Enquire",
                        "title": "Is paid",
                        "description": "Is an enquire paid",
                        "format": "boolean",
                        "example": "true"
                    },
                    "location": {
                        "schema": "Enquire",
                        "$ref": "#/components/schemas/LocationResource"
                    },
                    "answers": {
                        "schema": "Enquire",
                        "title": "Answers",
                        "description": "An enquire's answers",
                        "items": {
                            "type": "object",
                            "$ref": "#/components/schemas/EnquireAnswerResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "LocationResource": {
                "title": "Location resource",
                "description": "Resource for a location representation",
                "properties": {
                    "id": {
                        "schema": "Location",
                        "title": "ID",
                        "description": "A locations's identificator",
                        "format": "int64",
                        "example": "5"
                    },
                    "country": {
                        "schema": "Location",
                        "title": "A location's country",
                        "description": "A location's country",
                        "format": "string",
                        "example": "USA"
                    },
                    "city": {
                        "schema": "Location",
                        "title": "A location's city",
                        "description": "A location's city",
                        "format": "string",
                        "example": "New York"
                    },
                    "state": {
                        "schema": "Location",
                        "title": "A location's state",
                        "description": "A location's state",
                        "format": "string",
                        "example": "New York"
                    },
                    "postal_code": {
                        "schema": "Location",
                        "title": "A location's postal code",
                        "description": "A location's postal code",
                        "format": "string",
                        "example": "12345"
                    },
                    "address": {
                        "schema": "Location",
                        "title": "A location's address",
                        "description": "A location's address",
                        "format": "string",
                        "example": "address"
                    },
                    "latitude": {
                        "schema": "Location",
                        "title": "A location's latitude",
                        "description": "A location's latitude",
                        "format": "double",
                        "example": 5.123,
                        "nullable": true
                    },
                    "longitude": {
                        "schema": "Location",
                        "title": "A location's longitude",
                        "description": "A location's longitude",
                        "format": "double",
                        "example": 8.123,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    }
}